[Bronze I] Title: 평균, Time: 256 ms, Memory: 18788 KB -BaekjoonHub

+++ 백준/Bronze/1546. 평균/평균.java
@@ -5,7 +5,7 @@ public class Main {
         Scanner sc = new Scanner(System.in);
         int examCount = sc.nextInt();
         int[] examScores = new int[examCount];
-        double sum = 0;
+        int sum = 0;
 
         for (int i = 0; i < examCount; i++) {
             examScores[i] = sc.nextInt();
@@ -14,6 +14,6 @@ public class Main {
 
         int max = Arrays.stream(examScores).max().orElseThrow();
 
-        System.out.println(sum * 100 / max / examCount);
+        System.out.println(sum * 100.0 / max / examCount);
     }
 }
